Feature added:
*Added code to rotate the bitmap to look correct if rotated incorrectly
 rotateImage() function:

This code is a function that rotates a `Bitmap` image by a specified number of degrees. It takes two arguments: the `Bitmap` image to rotate (`img`), and the number of degrees to rotate it by (`degree`). It returns a new `Bitmap` object that represents the rotated image.

Here's a step-by-step breakdown of what the code does:

1. The function creates a new `Matrix` object called `matrix`. A `Matrix` is a mathematical object that can be used to transform (i.e. rotate, scale, translate, etc.) a `Bitmap` image.

2. The function calls the `postRotate()` method on the `matrix` object, passing in the `degree` argument. This rotates the `matrix` by the specified number of degrees.

3. The function calls the `createBitmap()` method on the `Bitmap` object (`img`), passing in the `matrix` object, as well as the x and y coordinates of the top-left corner of the image (`0, 0`), the width and height of the image (`img.width` and `img.height`, respectively), and a boolean value that specifies whether the new `Bitmap` should be mutable (`true`). This creates a new `Bitmap` object that represents the rotated image.

4. The function calls the `recycle()` method on the original `Bitmap` object (`img`). This frees up memory that was used by the original `Bitmap` object.

5. The function returns the new `Bitmap` object (`rotatedImg`), which represents the rotated image.

In summary, this code rotates a `Bitmap` image by a specified number of degrees using a `Matrix` object. It creates a new `Bitmap` object that represents the rotated image, and frees up memory used by the original `Bitmap` object.

* rotateImageIfRequired() function:
This code is a function that checks if a `Bitmap` image needs to be rotated based on its metadata, and rotates it if necessary. It takes three arguments: a `Context` object (`context`), a `Bitmap` image (`img`), and a `Uri` object that represents the location of the image (`selectedImage`). It returns a new `Bitmap` object that represents the rotated image.

Here's a step-by-step breakdown of what the code does:

1. The function uses the `contentResolver` property of the `Context` object to open an input stream to the image file specified by the `selectedImage` argument. It also gets the path to the image file.

2. The function creates a new `ExifInterface` object called `ei`. An `ExifInterface` is a class that can be used to read metadata from image files. The `when` expression is used to create a conditional statement that checks the version of the Android operating system and whether the input stream is not null. If the version is greater than 23 and the input stream is not null, the `ExifInterface` is created using the input stream. Otherwise, if the path is not null, the `ExifInterface` is created using the path. If neither condition is true, the function returns the original `Bitmap` object (`img`).

3. The function uses the `getAttributeInt()` method of the `ExifInterface` object to get the orientation of the image. The `ExifInterface.TAG_ORIENTATION` constant is used to specify the orientation tag. The `getAttributeInt()` method returns an integer value that represents the orientation of the image.

4. The function uses a `when` expression to check the orientation value and rotate the image if necessary. If the orientation is `ExifInterface.ORIENTATION_ROTATE_90`, the `rotateImage()` function is called with a rotation angle of 90 degrees. If the orientation is `ExifInterface.ORIENTATION_ROTATE_180`, the `rotateImage()` function is called with a rotation angle of 180 degrees. If the orientation is `ExifInterface.ORIENTATION_ROTATE_270`, the `rotateImage()` function is called with a rotation angle of 270 degrees. Otherwise, the function returns the original `Bitmap` object (`img`).

5. The `rotateImage()` function is a helper function that rotates a `Bitmap` image by a specified number of degrees, as explained in the previous code explanation.

In summary, this code checks if a `Bitmap` image needs to be rotated based on its metadata, and rotates it if necessary. It uses an `ExifInterface` object to read the orientation metadata, and a `when` expression to check the orientation value and rotate the image if necessary. It returns a new `Bitmap` object that represents the rotated image.

Next we add the setImage() method to the BookmarkDetailsViewModel.kt class