Feature: 1. added search feature.
2. Creating ad-hoc bookmarks:
What if the user wants
to add a bookmark for a place that doesn’t show up on the map? We can make this
possible by allowing the user to drop a pin at any location on the map.
Currently, MapsViewModel includes a method to create a bookmark from a place, but
now we need one to create a bookmark from only a map location. We call this method *addBookmark()* in the MapsViewModel.kt class.