Method: newBookmark()
After implementing the addBookmark() method in MapsViewModel.kt, we need a method in MapsActivity.kt to take advantage of addBookmark().

This method creates a new bookmark from a location, and then it starts the
bookmark details Activity to allow editing of the new bookmark. The call to
addBookmark runs within a coroutine block because it accesses the database and
can’t run on the main thread.